+++ src/enums/o_auth_provider.rs
@@ -89,6 +89,10 @@ pub enum OAuthProvider {
     Zoho,
     #[serde(rename = "zoom")]
     Zoom,
+    #[serde(rename = "githubImagine")]
+    GithubImagine,
+    #[serde(rename = "googleImagine")]
+    GoogleImagine,
 }
 
 impl OAuthProvider {
@@ -138,6 +142,8 @@ impl OAuthProvider {
             OAuthProvider::Yandex => "yandex",
             OAuthProvider::Zoho => "zoho",
             OAuthProvider::Zoom => "zoom",
+            OAuthProvider::GithubImagine => "githubImagine",
+            OAuthProvider::GoogleImagine => "googleImagine",
         }
     }
 }